requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
twilio==8.10.0
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
Flask-Login==0.6.3
Werkzeug==2.3.7
